Consolidate Categories and ReportCategories classes
