+++ testplan/exporters/testing/pdf/renderers/reports.py
@@ -7,9 +7,8 @@ from reportlab.lib import colors
 from testplan.common.exporters.pdf import RowStyle
 from testplan.common.utils.registry import Registry
 from testplan.common.utils.strings import format_description, wrap, split_text
-from testplan.testing.multitest.base import Categories
-from testplan.report.testing import (
-    Status, TestReport, TestGroupReport, TestCaseReport
+from testplan.report import (
+    Status, TestReport, TestGroupReport, TestCaseReport, ReportCategories
 )
 from testplan.testing import tagging
 from . import constants as const
@@ -282,9 +281,9 @@ class TestRowRenderer(BaseRowRenderer, MetadataMixin):
         Filter out passing rows if `failing_tests` is `True`.
         """
         style = self.get_style(source)
-        if source.category == Categories.SUITE:
+        if source.category == ReportCategories.SUITE:
             return style.display_suite
-        elif source.category == Categories.PARAMETRIZATION:
+        elif source.category == ReportCategories.PARAMETRIZATION:
             return style.display_case
         return style.display_test
 
@@ -307,7 +306,7 @@ class TestCaseRowBuilder(TestRowRenderer):
         return self.get_style(source).display_case
 
 
-@registry.bind((TestGroupReport, Categories.MULTITEST))
+@registry.bind((TestGroupReport, ReportCategories.MULTITEST))
 class MultiTestRowBuilder(TestRowRenderer):
     """Multitests get special treatment with extra formatting & summary."""
