+++ tests/functional/testplan/report/testing/test_merge.py
@@ -14,7 +14,7 @@ should be equal to the one that is generated by merging multiple ones.
 """
 
 from testplan.common.utils.testing import check_report
-from testplan.report.testing import TestCaseReport, TestGroupReport
+from testplan.report import TestCaseReport, TestGroupReport
 
 
 # TODO: Multitest/suite splitting is not supported yet,
