+++ tests/functional/testplan/testing/multitest/test_suite_decorators.py
@@ -5,9 +5,8 @@ from testplan.testing.multitest import MultiTest
 from testplan import Testplan
 from testplan.common.utils.testing import log_propagation_disabled
 from testplan.common.utils.logger import TESTPLAN_LOGGER
-from testplan.report.testing import (TestReport, TestGroupReport,
-                                     TestCaseReport)
-from testplan.testing.multitest.base import Categories
+from testplan.report import (TestReport, TestGroupReport,
+                             TestCaseReport, ReportCategories)
 from testplan.testing.multitest.suite import (testcase, testsuite, skip_if,
                                               post_testcase, pre_testcase)
 
@@ -105,7 +104,7 @@ def test_basic_multitest():
             for tc_entry in st_entry.entries:
                 if tc_entry.name == 'case4':
                     assert isinstance(tc_entry, TestGroupReport)
-                    assert tc_entry.category == Categories.PARAMETRIZATION
+                    assert tc_entry.category == ReportCategories.PARAMETRIZATION
                     assert len(tc_entry.entries) == 2  # 2 generated testcases
                 else:
                     assert isinstance(tc_entry, TestCaseReport)
