+++ tests/functional/testplan/testing/test_tagging.py
@@ -2,7 +2,7 @@ import pytest
 
 from testplan.common.utils.testing import check_report, log_propagation_disabled
 
-from testplan.report.testing import TestReport, TestGroupReport, TestCaseReport
+from testplan.report import TestReport, TestGroupReport, TestCaseReport
 from testplan.testing.multitest import MultiTest, testsuite, testcase
 from testplan.common.utils.logger import TESTPLAN_LOGGER
